refactor: split GameService resume, turn-trace, and event-handler lifecycle (#4582)
Move diplomacy resume wrappers, turn-trace order aggregation, and dialog builders into sibling libraries so AC7 hosts stay ≤240 lines without changing resume or persist semantics.
