+++ app/lib/core/services/app_event_handler/app_event_handler_scope_state.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import 'app_event_handler_scope.dart';
+import 'app_event_handler_scope_dialog_builders.dart';
 import 'app_event_handler_scope_lifecycle.dart';
 import 'app_event_handler_scope_session_helpers.dart';
 import 'app_event_handler_scope_session_subscriptions.dart';
